rtld-elf: compile with WANRS=4 warnings other than -Wcast-align
Reviewed By:	kib
Approved By:	brooks (mentor)
Differential Revision: https://reviews.freebsd.org/D17153

# Conflicts:
#	libexec/rtld-elf/map_object.c
#	libexec/rtld-elf/rtld.c

+++ libexec/rtld-elf/malloc.c
@@ -157,8 +157,8 @@ malloc(size_t nbytes)
 {
 	union overhead *op;
 	int bucket;
-	size_t n;
-	unsigned amt;
+	ssize_t n;
+	size_t amt;
 
 	/*
 	 * First time malloc is called, setup page size and
